+++ crates/integrations/connector-integration/src/connectors/nexinets.rs
@@ -636,6 +636,7 @@ macros::macro_connector_flow_status_impls!(
         IncrementalAuthorization,
         CreateOrder,
         CreateConnectorCustomer,
+        GetConnectorCustomer,
         MandateRevoke,
         PaymentMethodToken,
         PreAuthenticate,