+++ crates/types-traits/interfaces/src/connector_types.rs
@@ -97,6 +97,7 @@ pub trait ConnectorServiceTrait<T: PaymentMethodDataTypes>:
     + ServerSessionAuthentication
     + ServerAuthentication
     + CreateConnectorCustomer
+    + GetConnectorCustomer
     + PaymentTokenV2<T>
     + RechargeV2
     + CreatePaymentMethodV2
@@ -210,6 +211,10 @@ pub trait ValidationTrait: ConnectorCommon {
         false
     }
 
+    fn should_get_connector_customer(&self) -> bool {
+        false
+    }
+
     fn should_do_payment_method_token(
         &self,
         _payment_method: PaymentMethod,
@@ -296,6 +301,16 @@ pub trait ServerAuthentication:
 pub trait CreateConnectorCustomer:
     ConnectorIntegrationV2<
     connector_flow::CreateConnectorCustomer,
+    PaymentFlowData,
+    ConnectorCustomerData,
+    ConnectorCustomerResponse,
+>
+{
+}
+
+pub trait GetConnectorCustomer:
+    ConnectorIntegrationV2<
+    connector_flow::GetConnectorCustomer,
     PaymentFlowData,
     ConnectorCustomerData,
     ConnectorCustomerResponse,